% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deter_cutoff.R
\encoding{UTF-8}
\name{deter_cutoff}
\alias{deter_cutoff}
\title{Threshold's optimization}
\usage{
deter_cutoff(
  alpha = 0.1,
  n_bras = 1,
  bonf = FALSE,
  nsim_oc = 10000,
  ana_inter,
  ana_inter_tox = NULL,
  rand_ratio = NULL,
  p_n,
  prior = NULL,
  p_a,
  phi = NULL,
  delta = NULL,
  mat_beta_xi = matrix(c(1, 1, 0, 0, 0, 1, 0, 1), nrow = 2, byrow = TRUE),
  cut_seq = seq(0.5, 1, by = 0.005),
  power_seq = seq(0, 1, by = 0.01),
  seed = 1024,
  methode = 4L,
  affich_mat = NULL
)
}
\arguments{
\item{alpha}{Maximal type I error rate.}

\item{n_bras}{Number of treatment arms.}

\item{bonf}{TRUE to apply a Bonferroni correction.}

\item{nsim_oc}{Number of simulated trials.}

\item{ana_inter}{Vector giving the number of additional patients at each interim analysis. If ana_inter_tox is supplied, represents the vector for efficacy interim analyses.}

\item{ana_inter_tox}{Vector of the number of additional patients at each interim analysis for toxicity. If analyses for effficacy and toxicity occur at the same
number of patients, set it to NULL.}

\item{rand_ratio}{Number or numeric vector representing the ratio between the number of patients in treatment arms and control arm (or some reference sample size in case
of comparison to a reference value).}

\item{p_n}{Law of probability under the inefficacy/toxicity hypothesis.}

\item{prior}{Prior for the law of probability (by default, skeptical: the inefficacy/toxicity hypothesis).}

\item{p_a}{Law of probability under the efficacy/non toxicity hypothesis.}

\item{phi}{Thresholds for efficacy and toxicity (by default the values under inefficacy/toxicity hypothesis). If specified, the values for efficacy and non toxicity.}

\item{delta}{Minimal desirable difference of effect if compared with a control arm. Should be between -1 and 1.}

\item{mat_beta_xi}{The link matrix between outcomes of the multinomial distribution and endpoints. (Second row is for non toxicity instead of toxicity).}

\item{cut_seq, power_seq}{Vectors of values used to choose the couple while optimizing the Cn.}

\item{seed}{Seed for \code{rmultinom}.}

\item{methode}{Method used to generate the data:
\itemize{
\item 1: trial by trial;
\item 2: trial by trial, patient by patient;
\item 3: whole in 1;
\item 4: same generation as method 2, but written in C++ for faster computations.
}}

\item{affich_mat}{Yes to display in console the matrix used for the choice, No for no display and default option NULL will ask you.}
}
\value{
Displaying in console of the result.
If stocked in an object, a list:
\itemize{
\item the optimal couple λ/γ and α-risk and power a priori;
\item the matrix of the choices (not available in C++).
}
}
\description{
Optimization of λ and γ in threshold Cn = 1 - λ * (n / N) ^ γ.
}
\details{
2 situations:
\itemize{
\item for 1 arm: screening of the couples λ/γ by simulating monoarm trials under inefficacy/toxicity hypothesis to assess the control of α risk; and under
the efficacy/non toxicity hypothesis to assess the power;
\item for multiple arms with Bonferroni correction: same as for 1 arm with a risk of α/n_bras;
\item for multiple arms without Bonferroni correction: screening of the couples λ/γ by simulating multiarm trials under Global Null Hypothesis to assess the control of FWER; and under
the Least Favourable Configuration to assess the power.
}
}
\examples{
# The trial mimicked in the article of Zhou et al., 2017
carac_opt <- deter_cutoff(alpha = 0.1,
                          ana_inter = c(10, rep(5, 6)),
                          p_n = c(0.15, 0.3, 0.15, 0.4),
                          p_a = c(0.18, 0.42, 0.02, 0.38),
                          methode = 3L,
                          affich_mat = "No")
carac_opt[[1]]

# A trial with 3 arms
carac_opt <- deter_cutoff(alpha = 0.1,
                          ana_inter = rep(15, 4),
                          p_n = c(0.1, 0.2, 0.2, 0.5),
                          p_a = c(0.15, 0.5, 0.05, 0.3),
                          n_bras = 3,
                          affich_mat = "No",
                          bonf = FALSE)
carac_opt[[1]]

}
\references{
Zhou, H., Lee, J. J., & Yuan, Y. (2017). BOP2: Bayesian optimal design for phase II clinical trials with simple and complex endpoints. Statistics in Medicine, 36: 3302-3314.

Lin, R., Coleman, R. L., and Yuan, Y. (2020). TOP: Time-to-event Bayesian optimal phase II trial design for cancer immunotherapy. Journal of the National Cancer Institute 112, 38-45.
}
