% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_patients.R
\encoding{UTF-8}
\name{gen_patients_multinom}
\alias{gen_patients_multinom}
\title{Generate data}
\usage{
gen_patients_multinom(
  n_sim,
  ana_inter,
  ana_inter_tox = NULL,
  rand_ratio = NULL,
  multinom_cont = NULL,
  multinom_ttt = list(),
  mat_beta_xi = matrix(c(1, 1, 0, 0, 0, 1, 0, 1), nrow = 2, byrow = TRUE),
  seed = 1024,
  methode = 2L
)
}
\arguments{
\item{n_sim}{The desired number of simulated trials.}

\item{ana_inter}{Vector giving the number of additional patients at each interim analysis. If ana_inter_tox is supplied, represents the vector for efficacy interim analyses.
Else both efficacy and toxicity analyses are considered at the same time.}

\item{ana_inter_tox}{Vector of the number of additional patients at each interim analysis for toxicity. If analyses for effficacy and toxicity occur at the same
number of patients, set it to NULL.}

\item{rand_ratio}{Number or numeric vector representing the ratio between the number of patients in treatment arms and control arm (or some reference sample size in case
of comparison to a reference value).}

\item{multinom_cont}{Vector of length 4 describing the multinomial distribution followed by the control group ("Eff/Tox", "Eff/NoTox", "NoEff/Tox", "NoEff/NoTox").
If no control group, let it set to default NULL.}

\item{multinom_ttt}{A list of vectors of length 4 describing the multinomial law followed by each treatment group ("Eff/Tox", "Eff/NoTox", "NoEff/Tox", "NoEff/NoTox").
If only one treatment group, one can just specify a vector of length 4.}

\item{mat_beta_xi}{The link matrix between outcomes of the multinomial distribution and endpoints. (Second row is for non toxicity instead of toxicity).}

\item{seed}{The seed for \code{rmultinom}.}

\item{methode}{Method used to generate the data:
\itemize{
\item 1: trial by trial;
\item 2: trial by trial, patient by patient;
\item 3: whole in 1;
\item 4: like method 2, but in C++ to be faster.
}}
}
\value{
A data.frame with 1 row by interim analysis and treatment group, and 10 variables:
\itemize{
\item n_sim: the index of the trial;
\item nb_ana: the index of the interim analysis;
\item ttt: the treatment arm;
\item efftox, effnotox, noefftox, noeffnotox: cumulative number of each outcome at each interim analysis for each trial;
\item tot_pat, tot_eff, tot_notox: the total of patients, responses and non toxicities at each interim analysis for each trial.
}
}
\description{
Generate the data of \code{n_sim} trials.
}
\details{
Uses the multinomial law to simulate \code{n_sim} trials with \code{rmultinom}.

With or without control group, and with 1 final analysis or possible multiple interim analyses.

Data are in long format.

By default we chose method 2 because it seemed to be less biased (but all method are very close).
}
\examples{
# 100 trials of 40 patients with an analysis after each enrollment of 10 patients. Only 1 treatment group.
gen_patients_multinom(n_sim = 100,
                      ana_inter = rep(10, 4),
                      multinom_ttt = rep(0.25, 4))

# 100 trials of 40 patients with an analysis after each enrollment of 10 patients. 3 treatment groups.
gen_patients_multinom(n_sim = 100,
                      ana_inter = rep(10, 4),
                      multinom_ttt = list(rep(0.25, 4), c(0.2, 0.4, 0.2, 0.2), c(0.1, 0.4, 0.05, 0.45)))

# 100 trials of 40 patients with an analysis after each enrollment of 10 patients. 1 treamtent group with 1 control group.
gen_patients_multinom(n_sim = 100,
                      ana_inter = rep(10, 4),
                      multinom_ttt = rep(0.25, 4),
                      multinom_cont = c(0.2, 0.3, 0.1, 0.4))

# 100 trials of 40 patients with an analysis after each enrollment of 10 patients. 3 treatment groups with a control group.
gen_patients_multinom(n_sim = 100,
                      ana_inter = rep(10, 4),
                      multinom_ttt = list(rep(0.25, 4), c(0.2, 0.4, 0.2, 0.2), c(0.1, 0.4, 0.05, 0.45)),
                      multinom_cont = c(0.2, 0.3, 0.1, 0.4))

# 100 trials of 40 patients with an analysis after each enrollment of 10 patients. 3 treatment groups with a
# 40 patients control group. Each group has a different number of patients
# (for the 0.75 ratio, 0.75 * 10 = 7.5 rounded to 8 additional patients at each interim analysis).
gen_patients_multinom(n_sim = 100,
                      rand_ratio = c(1, .75, 1.5),
                      ana_inter = rep(10, 4),
                      multinom_ttt = list(rep(0.25, 4), c(0.2, 0.4, 0.2, 0.2), c(0.1, 0.4, 0.05, 0.45)),
                      multinom_cont = c(0.2, 0.3, 0.1, 0.4))
}
