% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_patients_TOP.R
\encoding{UTF-8}
\name{gen_temps_patient}
\alias{gen_temps_patient}
\title{Generate a time of entry or of event}
\usage{
gen_temps_patient(
  loi = c("uniform", "weibull", "log-logistic"),
  n = 1L,
  tmax,
  ptmax = 1,
  prop_demi = 0.5
)
}
\arguments{
\item{loi}{The distribution law (uniform, weibull or log-logistic).}

\item{n}{An positive integer giving the number of patients.}

\item{tmax}{A positive real giving the maximum time of the observation window.}

\item{ptmax}{For Weibull and log-logistic distributions, the proportion of patients with endpoint
at \code{tmax}. (But all n times will be between 0 and \code{tmax}. It is used to determine the 2
parameters of the distribution).}

\item{prop_demi}{For Weibull and log-logistic distributions, the proportion of \code{ptmax} at \code{tmax}/2.}
}
\value{
A vector of \code{n} times between 0 and \code{tmax} simulated according to the wanted distribution.
}
\description{
The function generates times for patients with the specified law. There are actually
3 laws supported : uniform, Weibull and log-logistic.
It is assumed that the time is between 0 and \code{tmax}. With uniform distribution, 50\% of the patients will have
a time between 0 and \code{tmax}/2; and with Weibull/log-logistic distributions, \code{prop_demi}\% of
the patients will have a time between 0 and \code{tmax}/2.
}
\examples{
# Generate the time of arrival of 10 successive patients with a uniform
# recruitment of mean 30 days
time_pat <- gen_temps_patient(loi = "unif",
                              n = 10L,
                              tmax = 2 * 30)
cumsum(time_pat)

# Generate the time of efficacy of 20 patients after their recruitment
# with a Weibull distribution between 0 and 100 days with 75\% of the responses
# in the first half of the window.
gen_temps_patient(loi = "weib",
                  n = 20L,
                  tmax = 100, ptmax = .8,
                  prop_demi = .75)
}
