% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deter_cutoff.R
\encoding{UTF-8}
\name{get_stopbound}
\alias{get_stopbound}
\title{Table of the stopping boundaries}
\usage{
get_stopbound(
  ana_inter,
  ana_inter_tox = NULL,
  rand_ratio = NULL,
  C_,
  gamm,
  p_n,
  prior = NULL,
  mat_beta_xi = matrix(c(1, 1, 0, 0, 0, 1, 0, 1), nrow = 2, byrow = TRUE),
  phi = NULL,
  delta = NULL
)
}
\arguments{
\item{ana_inter}{Vector giving the number of additional patients at each interim analysis. If ana_inter_tox is supplied, represents the vector for efficacy interim analyses.}

\item{ana_inter_tox}{Vector of the number of additional patients at each interim analysis for toxicity. If analyses for effficacy and toxicity occur at the same
number of patients, set it to NULL.}

\item{rand_ratio}{Number or numeric vector representing the ratio between the number of patients in treatment arms and control arm (or some reference sample size in case
of comparison to a reference value).}

\item{C_, gamm}{The 2 tuning parameters of criterion Cn.}

\item{p_n}{Law of probability under the inefficacy/toxicity hypothesis.}

\item{prior}{Prior for the law of probability (by default, skeptical: the inefficacy/toxicity hypothesis).}

\item{mat_beta_xi}{The link matrix between outcomes of the multinomial distribution and endpoints (efficacy and non toxicity /!\).}

\item{phi}{Thresholds for efficacy and toxicity (by default the values under inefficacy/toxicity hypothesis). If specified, the values for efficacy and non toxicity.}

\item{delta}{Minimal desirable difference of effect if compared with a control arm. If no control group, leave NULL. Should be between -1 and 1.}
}
\description{
Given a couple λ/γ, give the table of stopping boundaries at each interim analysis.
}
\details{
Generate the stopping boundaries table different given the situation:
\itemize{
\item against a fixed reference value: a data.frame with thresholds at each interim analysis. nb_ana give the number of patients at each
interim analysis, ttt the treatment group and tot_pat the number of recruited patients. seuil_eff is the threshold for futility (stop for futility
if the number of responses is less or equal to this threshold), seuil_notox and seuil_tox for toxicity (stop for toxicity if the number of non toxicities
is less or equal to seuil_notox or if the number of toxicities is greater or equal to seuil_tox).
\item against a control treatment: a data.frame with thresholds at each interim analysis with the values for each possible number of responses/non toxicities
in the control group. seuil_eff and seuil_notox works in the same way as above.
}
}
\examples{
# The table from the article of Zhou et al. de 2017
get_stopbound(ana_inter = c(10, rep(5, 6)),
              C_ = 0.625, gamm = 0.98,
              p_n = c(0.15, 0.3, 0.15, 0.4))

# The table for a larger trial with 3 arms
get_stopbound(ana_inter = rep(15, 4),
              C_ = 0.78, gamm = 0.9,
              p_n = c(0.15, 0.3, 0.15, 0.4))

}
