% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deter_constcutoff.R
\encoding{UTF-8}
\name{get_stopbound_const}
\alias{get_stopbound_const}
\title{Stopping boundaries for constant threshold}
\usage{
get_stopbound_const(
  ana_inter,
  rand_ratio = NULL,
  seuils,
  p_n,
  prior = NULL,
  mat_beta_xi = matrix(c(1, 1, 0, 0, 0, 1, 0, 1), nrow = 2, byrow = TRUE),
  phi = NULL,
  delta = NULL
)
}
\arguments{
\item{ana_inter}{Vector giving the number of additional patients at each interim analysis.}

\item{rand_ratio}{Number or numeric vector representing the ratio between the number of patients in treatment arms and control arm (or some reference sample size in case
of comparison to a reference value). 1 by default for all groups.}

\item{seuils}{The 2 constant probability thresholds for efficacy and toxicity.}

\item{p_n}{Law of probability under the inefficacy/toxicity hypothesis.}

\item{prior}{Prior for the law of probability (by default, skeptical: the inefficacy/toxicity hypothesis).}

\item{mat_beta_xi}{The link matrix between outcomes of the multinomial distribution and endpoints (efficacy and non toxicity /!\).}

\item{phi}{Thresholds for efficacy and toxicity (by default the values under inefficacy/toxicity hypothesis). If specified, the values for efficacy and non toxicity.}

\item{delta}{Minimal desirable difference of effect if compared with a control arm. If no control group, leave NULL. Should be between -1 and 1.}
}
\description{
Get the table of stopping boundaries for a given couple of thresholds.
}
\details{
Generation of a table for stopping boundaries. It can be:
\itemize{
\item for a design vs a reference value: a data.frame with the number of patients at each interim analysis with efficacy's thresholds (stop for futility if
less or equal responses than the threshold), non toxicity's thresholds (stop for toxicity if less or equal non toxicities than the threshold) and toxicity's thresholds
(stop for toxicity if more or equal toxicities than this threshold). If all groups aren't of the same size for all treatments, data in long format with a column
ttt indicating what treatment group the tresholds are for.
\item for a design vs a control group: a data.frame with the efficacy's thresholds (stop for futility if less or equal responses than the threshold when there are
y_cont responses in the control group) and non toxicity's threshold (stop for toxicity if less or equal non toxicities than the threshold when there are
y_cont non toxicities in the control group). If all groups aren't of the same size for all treatments, data in long format with a column
ttt indicating what treatment group the tresholds are for.
}
}
\examples{
# To get the table of stopping boundaries vs a given reference value
get_stopbound_const(ana_inter = rep(10, 4),
                    seuils = c(.67, .39),
                    p_n = c(.15, .3, .15, .4))

}
