% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deter_constcutoff.R
\name{getoc_tox_const}
\alias{getoc_tox_const}
\title{Get the operating characteristics for a given couple gamma_eff/gamma_tox et a given list of trials.}
\usage{
getoc_tox_const(
  ana_inter,
  rand_ratio = NULL,
  prior = NULL,
  p_n,
  seuils,
  mat_beta_xi = matrix(c(1, 1, 0, 0, 0, 1, 0, 1), nrow = 2, byrow = TRUE),
  phi = NULL,
  delta = c(0, 0),
  liste_patients
)
}
\arguments{
\item{ana_inter}{Vector giving the number of additional patients at each interim analysis.}

\item{rand_ratio}{Number or numeric vector representing the ratio between the number of patients in treatment arms and control arm (or some reference sample size in case
of comparison to a reference value). 1 by default for all groups.}

\item{prior}{Prior for the law of probability (by default, skeptical: the inefficacy/toxicity hypothesis).}

\item{seuils}{The 2 constant probability thresholds for efficacy and toxicity.}

\item{mat_beta_xi}{The link matrix between outcomes of the multinomial distribution and endpoints (efficacy and non toxicity /!\).}

\item{phi}{Thresholds for efficacy and toxicity (by default the values under inefficacy/toxicity hypothesis). If specified, the values for efficacy and non toxicity.}

\item{delta}{Minimal desirable difference of effect if compared with a control arm. If no control group, leave NULL. Should be between -1 and 1.}

\item{liste_patients}{Data.frame of generated trials with \code{gen_patients_multinom}.}

\item{n_sim}{Number of simulated trials.}
}
\description{
Get the operating characteristics for a given couple gamma_eff/gamma_tox et a given list of trials.
}
