% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deter_cutoff_holm.R
\encoding{UTF-8}
\name{getoc_tox_holm2}
\alias{getoc_tox_holm2}
\title{Operating characteristics for a couple &lambda_holm;/&gamma_holm; given a trials' list and the same hyperparameters for a monoarm trial.}
\usage{
getoc_tox_holm2(
  ana_inter,
  ana_inter_tox = NULL,
  rand_ratio = NULL,
  prior = NULL,
  p_n,
  m,
  C_holm,
  gamm_holm,
  C_mono,
  gamm_mono,
  mat_beta_xi = matrix(c(1, 1, 0, 0, 0, 1, 0, 1), nrow = 2, byrow = TRUE),
  phi = NULL,
  delta = c(0, 0),
  liste_patients
)
}
\arguments{
\item{ana_inter}{Vector giving the number of additional patients at each interim analysis. If ana_inter_tox is supplied, represents the vector for efficacy interim analyses.}

\item{ana_inter_tox}{Vector of the number of additional patients at each interim analysis for toxicity. If analyses for effficacy and toxicity occur at the same
number of patients, set it to NULL.}

\item{rand_ratio}{Number or numeric vector representing the ratio between the number of patients in treatment arms and control arm (or some reference sample size in case
of comparison to a reference value). Not yet implemented.}

\item{prior}{Prior for the law of probability (by default, skeptical: the inefficacy/toxicity hypothesis).}

\item{p_n}{Law of probability under the inefficacy/toxicity hypothesis.}

\item{m}{The maximum number of active arms.}

\item{C_holm, gamm_holm}{The 2 tuning parameters of criterion Cn optimized with the correction dependant on the number of active arms.}

\item{C_mono, gamm_mono}{The 2 tuning parameters of criterion Cn optimized for a single arm design.}

\item{mat_beta_xi}{The link matrix between outcomes of the multinomial distribution and endpoints (efficacy and non toxicity /!\).}

\item{phi}{Thresholds for efficacy and toxicity (by default the values under inefficacy/toxicity hypothesis). If specified, the values for efficacy and non toxicity.}

\item{delta}{Minimal desirable difference of effect if compared with a control arm. Should be between -1 and 1.}

\item{liste_patients}{Generated trials via function \code{gen_patients_multinom}.}
}
\description{
Operating characteristics for a couple &lambda_holm;/&gamma_holm; given a trials' list and the same hyperparameters for a monoarm trial.
}
