% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deter_cutoff.R
\name{max_resp}
\alias{max_resp}
\title{Get the boundaries for 1 number of patients. Give the maximum number of patients (or TRUE/FALSE in case of controlled design)
that satisfy Pr(p > phi | Dn) < 1 - Cn.}
\usage{
max_resp(
  cutoff,
  prior,
  ntr,
  ntr_cont = NULL,
  phi,
  delta = c(0, 0),
  mat_beta_xi = matrix(c(1, 1, 0, 0, 0, 1, 0, 1), nrow = 2, byrow = TRUE)
)
}
\arguments{
\item{cutoff}{Cn (in reality we give 1 - Cn).}

\item{prior}{The prior's law (by default the law under the inefficacy/toxicity hypothesis).}

\item{ntr}{Number of patients in treatment arm.}

\item{ntr_cont}{Number of patients in control arm.}

\item{phi}{Thresholds for efficacy and toxicity (by default the values under inefficacy/toxicity hypothesis).}

\item{delta}{Minimum desired difference for effect of treatment (by default 0 for efficacy and toxicity).}

\item{mat_beta_xi}{The link matrix between outcomes of the multinomial distribution and endpoints.}
}
\description{
Get the boundaries for 1 number of patients. Give the maximum number of patients (or TRUE/FALSE in case of controlled design)
that satisfy Pr(p > phi | Dn) < 1 - Cn.
}
