% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deter_constcutoff.R
\name{max_resp_const}
\alias{max_resp_const}
\title{Return the max number of responses/non toxicities before being greater than cutoff.}
\usage{
max_resp_const(
  cutoff,
  prior,
  ntr,
  ntr_cont = NULL,
  phi,
  delta = c(0, 0),
  mat_beta_xi = matrix(c(1, 1, 0, 0, 0, 1, 0, 1), nrow = 2, byrow = TRUE)
)
}
\arguments{
\item{cutoff}{Thresholds for efficacy and toxicity.}

\item{prior}{Lax of prior (by default, under H0).}

\item{ntr}{Number of enrolled patients in 1 arm.}

\item{phi}{Theoretical thresholds to which we compare the efficacy and toxicity (by default, values under H0).}

\item{mat_beta_xi}{Link matrix between endpoints and outcomes.}
}
\description{
Return the max number of responses/non toxicities before being greater than cutoff.
}
