% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_patients.R
\name{modif_tab_patients}
\alias{modif_tab_patients}
\title{Resample the interim analyses}
\usage{
modif_tab_patients(tab_patients, nouv_ana_inter)
}
\arguments{
\item{tab_patients}{Dataset generated with \code{gen_patients_multinom} with ana_inter = rep(1, X).}

\item{nouv_ana_inter}{The new vector of additional patients at each interim analysis. If several treatments,
you can either give 1 vector that will be applied to each group or give a nammed list with names corresponding to treatments.}
}
\description{
Resample the interim analyses based on the table with interim analyses after each patients.
}
\examples{
tab1 <- gen_patients_multinom(n_sim = 100,
                              ana_inter = rep(1, 40),
                              multinom_ttt = list(rep(0.25, 4), c(0.2, 0.4, 0.2, 0.4), c(0.1, 0.4, 0.05, 0.45))
tab2 <- modif_tab_patients(tab_patients = tab1,
                           nouv_ana_inter = rep(10, 4))
}
