% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opcharac_efftox.R
\encoding{UTF-8}
\name{opcharac_efftox}
\alias{opcharac_efftox}
\title{Operating characteristics}
\usage{
opcharac_efftox(
  alpha = 0.1,
  bonf = FALSE,
  nsim_oc = 10000,
  nsim_essai = 10000,
  ana_inter,
  ana_inter_tox = NULL,
  rand_ratio = 1,
  p_n,
  prior = NULL,
  p_a,
  p_reel,
  phi = NULL,
  delta = NULL,
  mat_beta_xi = matrix(c(1, 1, 0, 0, 0, 1, 0, 1), nrow = 2, byrow = TRUE),
  cut_seq = seq(0.5, 0.95, by = 0.005),
  power_seq = seq(0, 1, by = 0.01),
  seed_cut = 1024,
  seed_pat = 1993,
  methode = 2L,
  tableau_essais = NULL
)
}
\arguments{
\item{alpha}{Maximal type I error rate (FWER in case of multiarm).}

\item{bonf}{TRUE to apply a Bonferroni correction in case of multiarm settings.}

\item{nsim_oc, nsim_essai}{Number of simulated trials to optimize CN and get the operating characteristics.}

\item{ana_inter}{Vector giving the number of additional patients at each interim analysis. If ana_inter_tox is supplied, represents the vector for efficacy interim analyses.}

\item{ana_inter_tox}{Vector of the number of additional patients at each interim analysis for toxicity. If analyses for effficacy and toxicity occur at the same
number of patients, set it to NULL.}

\item{rand_ratio}{Number or numeric vector representing the ratio between the number of patients in treatment arms and control arm (or some reference sample size in case
of comparison to a reference value). Not yet implemented.}

\item{p_reel, p_n, prior, p_a}{Vectors of length 4 giving the law of probability under the 'real' scenario, the inefficacy/toxicity hypothesis, the prior
(by default \code{p_n}) and the efficacy/non toxicity hypothesis.}

\item{phi}{Thresholds for efficacy and toxicity (by default the values under inefficacy/toxicity hypothesis). If specified, the values for efficacy and non toxicity.}

\item{delta}{Minimal desirable difference of effect if compared with a control arm. Should be between -1 and 1.}

\item{mat_beta_xi}{The link matrix between outcomes of the multinomial distribution and endpoints. (Second row is for non toxicity instead of toxicity).}

\item{cut_seq, power_seq}{Vectors of values used to choose the couple while optimizing the Cn. If a unique value is specified for each parameter, the grid search won't be performed.}

\item{seed_cut, seed_pat}{Seeds for \code{rmultinom} (for Cn and operating characteristics).}

\item{methode}{Method used to generate the data:
\itemize{
\item 1: trial by trial;
\item 2: trial by trial, patient by patient;
\item 3: whole in 1.
}}

\item{tableau_essais}{Data.frame of simulated trials with \code{gen_patients_multinom}. If NULL, the function will simulated them using \code{p_reel}.}
}
\value{
\itemize{
\item for monoarm, a list:
\itemize{
\item caracteristique: data.frame with the global informations (type I error rate, inefficacy/toxicity and efficacy/non toxicity hypotheses,
simulation probabilities, method used to generate the data,
Cn, proportion of reject of inefficacy/toxicity hypothesis and early stoppings, early stoppings for futility/toxicity, mean number of stoppings (for
futility and toxicity), mean number of enrolled patients, responses and toxicities);
\item essais: data.frame of the simulated trials with 1 line for each trial (the final analysis).
}
\item for multiarm, a list:
\itemize{
\item carac_globales: data.frame with the global informations (FWER, inefficacy/toxicity and efficacy/non toxicity hypotheses,
simulation probabilities, method used to generate the data, Cn, proportion of global reject of inefficacy/toxicity hypothesis);
\item carac_bras: data.frame with the informations for each arm (proportion of reject of the inefficacy/toxicity hypothesis,
mean number of patients/responses/toxicities in each arm, proportion of early stoppings (and for futility/toxicity) and stoppings
(for futility and toxicity too));
\item essais: data.frame of the simulated trials.
}
}
}
\description{
Obtain the operating characteristics for BOP2 threshold vs a reference value or a control group.
}
\details{
Optimize Cn and then simulate \code{nsim_essais} with law \code{p_reel} to determine the operating characteristics. It is possible to optimize
Cn before with \code{deter_cutoff} and then supply these values as argument in \code{power_seq} and \code{cut_seq}. It allows reduction of the execution
if we want to simulate several scenarios with the same Cn.

Works for mono- and multiarm trials.

For multiarm settings, \code{bonf = TRUE} to perform Bonferroni correction (less time, but more conservative or even Cn).
}
\examples{
# For BOP2 simulated trials in Zhou, 2017
opcharac_efftox(ana_inter = c(10, rep(5, 6)),
                alpha = 0.1,
                p_reel = c(0.18, 0.42, 0.02, 0.38),
                p_n = c(0.15, 0.3, 0.15, 0.4), p_a = c(0.18, 0.42, 0.02, 0.38),
                nsim_oc = 10000, nsim_essai = 10000,
                seed_cut = 1024, seed_pat = 1024,
                methode = 3L)

# For a 3 arm-trial
opcharac_efftox(ana_inter = rep(15, 4),
                alpha = 0.1,
                p_reel = list(ttt1 = c(0.15, 0.5, 0.05, 0.3), ttt2 = c(0.15, 0.3, 0.25, 0.3), ttt3 = c(0.05, 0.15, 0.15, 0.65)),
                p_n = c(0.1, 0.2, 0.2, 0.5), p_a = c(0.15, 0.5, 0.05, 0.3),
                nsim_oc = 10000, nsim_essai = 10000)

}
