% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opcharac_efftox_const.R
\encoding{UTF-8}
\name{opcharac_efftox_const}
\alias{opcharac_efftox_const}
\title{Operating characteristics}
\usage{
opcharac_efftox_const(
  ana_inter,
  alpha = 0.1,
  mat_beta_xi = matrix(c(1, 1, 0, 0, 0, 1, 0, 1), nrow = 2, byrow = TRUE),
  phi = NULL,
  delta = NULL,
  p_reel,
  bonf = FALSE,
  p_n,
  p_a,
  nsim_oc = 10000,
  nsim_essai = 10000,
  tableau_essais = NULL,
  seq_eff = seq(0.01, 0.99, by = 0.02),
  seq_tox = seq(0.01, 0.99, by = 0.02),
  seed_cut = 1024,
  seed_pat = 1993,
  methode = 2L
)
}
\arguments{
\item{ana_inter}{Vector giving the number of additional patients at each interim analysis. If ana_inter_tox is supplied, represents the vector for efficacy interim analyses.}

\item{alpha}{Maximal type I error rate (FWER in case of multiarm).}

\item{mat_beta_xi}{The link matrix between outcomes of the multinomial distribution and endpoints. (Second row is for non toxicity instead of toxicity).}

\item{phi}{Thresholds for efficacy and toxicity (by default the values under inefficacy/toxicity hypothesis). If specified, the values for efficacy and non toxicity.}

\item{delta}{Minimal desirable difference of effect if compared with a control arm. Should be between -1 and 1.}

\item{p_reel, p_n, p_a}{Vectors of length 4 giving the law of probability under the 'real' scenario, the inefficacy/toxicity hypothesis and the efficacy/non toxicity hypothesis.}

\item{bonf}{TRUE to apply a Bonferroni correction in case of multiarm settings.}

\item{nsim_oc, nsim_essai}{Number of simulated trials to optimize CN and get the operating characteristics.}

\item{tableau_essais}{Data.frame of simulated trials with \code{gen_patients_multinom}.}

\item{seq_eff, seq_tox}{Vectors of values used to choose the couple while optimizing the couple gamma_eff/gamma_tox. If a unique value is specified for each parameter, the grid search won't be performed.}

\item{seed_cut, seed_pat}{Seeds for \code{rmultinom} (for Cn and operating characteristics).}

\item{methode}{Method used to generate the data:
\itemize{
\item 1: trial by trial;
\item 2: trial by trial, patient by patient;
\item 3: whole in 1.
}}
}
\value{
\itemize{
\item for monoarm, a list:
\itemize{
\item caracteristique: data.frame with the global informations (type I error rate, inefficacy/toxicity and efficacy/non toxicity hypotheses,
simulation probabilities, method used to generate the data,
Cn, proportion of reject of inefficacy/toxicity hypothesis and early stoppings, early stoppings for futility/toxicity, mean number of stoppings (for
futility and toxicity), mean number of enrolled patients, responses and toxicities);
\item essais: data.frame of the simulated trials with 1 line for each trial (the final analysis).
}
\item for multiarm, a list:
\itemize{
\item carac_globales: data.frame with the global informations (FWER, inefficacy/toxicity and efficacy/non toxicity hypotheses,
simulation probabilities, method used to generate the data, Cn, proportion of global reject of inefficacy/toxicity hypothesis);
\item carac_bras: data.frame with the informations for each arm (proportion of reject of the inefficacy/toxicity hypothesis,
mean number of patients/responses/toxicities in each arm, proportion of early stoppings (and for futility/toxicity) and stoppings
(for futility and toxicity too));
\item essais: data.frame of the simulated trials.
}
}
}
\description{
Obtain the operating characteristics with a constant threshold vs a reference value (or control group).
}
\details{
Optimize the couple gamma_eff/gamma_tox, then simulate \code{nsim_essais} trials under \code{p_reel} in order to determine the operating characteristics.
It is possible to optimise the couple before with the function \code{deter_constcutoff} and supply the obtained values in arguments
\code{seq_eff} and \code{seq_tox}.
The stopping rule can be written as stop if Pr(mat_beta_xi * p_ttt <= phi | Dn) > 1 - gamma_eff or 1 - gamma_tox.

Works for mono- and multiarm trials.

For multiarm settings, \code{bonf = TRUE} to perform Bonferroni correction (less time, but more conservative or even couple).
}
\examples{
# For BOP2 simulated trials in Zhou, 2017
opcharac_efftox_const(ana_inter = c(10, rep(5, 6)),
                      alpha = 0.1,
                      p_reel = c(0.18, 0.42, 0.02, 0.38),
                      p_n = c(0.15, 0.3, 0.15, 0.4), p_a = c(0.18, 0.42, 0.02, 0.38),
                      nsim_oc = 10000, nsim_essai = 10000,
                      seed_cut = 1024, seed_pat = 1993,
                      methode = 3L)

# For a 3 arm-trial
opcharac_efftox(ana_inter = rep(15, 4),
                alpha = 0.1,
                p_reel = list(ttt1 = c(0.15, 0.5, 0.05, 0.3), ttt2 = c(0.15, 0.3, 0.25, 0.3), ttt3 = c(0.05, 0.15, 0.15, 0.65)),
                p_n = c(0.1, 0.2, 0.2, 0.5), p_a = c(0.15, 0.5, 0.05, 0.3),
                nsim_oc = 10000, nsim_essai = 10000)

}
