% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opcharac_efftox_holm.R
\encoding{UTF-8}
\name{opcharac_efftox_holm2}
\alias{opcharac_efftox_holm2}
\title{Operating characteristics}
\usage{
opcharac_efftox_holm2(
  alpha = 0.1,
  nsim_oc = 10000,
  nsim_essai = 10000,
  ana_inter,
  ana_inter_tox = NULL,
  rand_ratio = 1,
  p_n,
  prior = NULL,
  p_a,
  p_reel,
  phi = NULL,
  delta = NULL,
  mat_beta_xi = matrix(c(1, 1, 0, 0, 0, 1, 0, 1), nrow = 2, byrow = TRUE),
  cut_seq_holm = seq(0.5, 0.95, by = 0.005),
  power_seq_holm = seq(0, 1, by = 0.01),
  cut_seq_mono = seq(0.5, 0.95, by = 0.005),
  power_seq_mono = seq(0, 1, by = 0.01),
  seed_cut = 1024,
  seed_pat = 1993,
  methode = 2L,
  tableau_essais = NULL
)
}
\arguments{
\item{alpha}{Maximal type I error rate (FWER in case of multiarm).}

\item{nsim_oc, nsim_essai}{Number of simulated trials to optimize CN and get the operating characteristics.}

\item{ana_inter}{Vector giving the number of additional patients at each interim analysis. If ana_inter_tox is supplied, represents the vector for efficacy interim analyses.}

\item{ana_inter_tox}{Vector of the number of additional patients at each interim analysis for toxicity. If analyses for effficacy and toxicity occur at the same
number of patients, set it to NULL.}

\item{rand_ratio}{Number or numeric vector representing the ratio between the number of patients in treatment arms and control arm (or some reference sample size in case
of comparison to a reference value). Not yet implemented.}

\item{p_reel, p_n, prior, p_a}{Vectors of length 4 giving the law of probability under the 'real' scenario, the inefficacy/toxicity hypothesis, the prior
(by default \code{p_n}) and the efficacy/non toxicity hypothesis.}

\item{phi}{Thresholds for efficacy and toxicity (by default the values under inefficacy/toxicity hypothesis). If specified, the values for efficacy and non toxicity.}

\item{delta}{Minimal desirable difference of effect if compared with a control arm. Should be between -1 and 1. Let to NULL if uncontrolled setting.}

\item{mat_beta_xi}{The link matrix between outcomes of the multinomial distribution and endpoints. (Second row is for non toxicity instead of toxicity).}

\item{cut_seq_holm, power_seq_holm}{Vectors of values used to choose the couple while optimizing the Cn. If a unique value is specified for each parameter, the grid search won't be performed.}

\item{cut_seq_mono, power_seq_mono}{Vectors of values used to choose the couple while optimizing the Cn for the monoarm trial. If a unique value is specified for each parameter, the grid search won't be performed.}

\item{seed_cut, seed_pat}{Seeds for \code{rmultinom} (for Cn and operating characteristics).}

\item{methode}{Method used to generate the data:
\itemize{
\item 1: trial by trial;
\item 2: trial by trial, patient by patient;
\item 3: whole in 1.
}}

\item{tableau_essais}{Data.frame of simulated trials with \code{gen_patients_multinom}.}
}
\value{
\itemize{
\item for monoarm, a list:
\itemize{
\item caracteristique: data.frame with the global informations (type I error rate, inefficacy/toxicity and efficacy/non toxicity hypotheses,
simulation probabilities, method used to generate the data,
Cn, proportion of reject of inefficacy/toxicity hypothesis and early stoppings, early stoppings for futility/toxicity, mean number of stoppings (for
futility and toxicity), mean number of enrolled patients, responses and toxicities);
\item essais: data.frame of the simulated trials with 1 line for each trial (the final analysis).
}
\item for multiarm, a list:
\itemize{
\item carac_globales: data.frame with the global informations (FWER, inefficacy/toxicity and efficacy/non toxicity hypotheses,
simulation probabilities, method used to generate the data, Cn, proportion of global reject of inefficacy/toxicity hypothesis);
\item carac_bras: data.frame with the informations for each arm (proportion of reject of the inefficacy/toxicity hypothesis,
mean number of patients/responses/toxicities in each arm, proportion of early stoppings (and for futility/toxicity) and stoppings
(for futility and toxicity too));
\item essais: data.frame of the simulated trials.
}
}
}
\description{
Obtain the operating characteristics for BOP2 like threshold vs a reference value or control group with correction in function of remaining active arms for cutoff determination.
}
\details{
Optimize Cn and then simulate \code{nsim_essais} with law \code{p_reel} to determine the operating characteristics. It is possible to optimize
Cn before with \code{deter_cutoff} and then supply these values as argument in \code{power_seq} and \code{cut_seq}. It allows reduction of the execution
if we want to simulate several scenarios with the same Cn.

The default form of Cn is 1 - (η - λ) / η * (n / N) ^ γ with η = K + 1 - k, K the number of arms and k the number of active arms.
At final analysis, the threshold is corrected with the monoarm threshold to maintain arm-specific type I error rate.
}
