% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opcharac_efftoxTOP.R
\name{realisation_essai}
\alias{realisation_essai}
\title{Running the ongoing of a trial with TOP}
\usage{
realisation_essai(
  ana_inter,
  ana_inter_tox = NULL,
  tableau,
  phitox,
  prior_eff,
  prior_tox,
  C_,
  gamm,
  interpatient,
  max_teff,
  max_ttox
)
}
\arguments{
\item{ana_inter}{A vector of the supplementary patients recruited at each interim analysis compared with the previous one.}

\item{ana_inter_tox}{Vector of the number of additional patients at each interim analysis for toxicity. If analyses for effficacy and toxicity occur at the same
number of patients, set it to NULL.}

\item{tableau}{The dataset of the trial (generated by \code{\link{gen_patients_multinomTOP}} or with similar structure).}

\item{phitox}{Theoretical thresholds to compare values of efficacy and toxicity.}

\item{prior_eff, prior_tox}{The prior for efficacy and toxicity.}

\item{C_, gamm}{The 2 parameters of the threshold Cn.}

\item{interpatient}{Mean time between 2 patients.}

\item{max_teff, max_ttox}{Length of the observation window for efficacy and toxicity.}
}
\value{
A data.frame of 1 row containing the following informations:
\itemize{
\item analyse: the index of interim analysis when the trial stopped;
\item nb_pts: the number of patients at the end of the trial;
\item temps_etude: the duration of the trial;
\item decision: the decision at the end of the trial;
\item arret_fut, arret_tox: 1 if the trial was stopped for futility/toxicity, 0 otherwise;
\item nb_eff, nb_tox: number of responses/toxicities at the time of end of the trial;
\item nb_eff_true, nb_tox_true: number of responses/toxicities at the time of end of the trial if all the patients had completed their observation window;
\item C_, Gamma: the 2 tuning parameters of the threshold Cn;
\item alpha_eff, beta_eff, alpha_tox, beta_tox: the 2 parameters of the beta distribution a posteriori for efficacy and toxicity at the end of the trial.
}
}
\description{
Given the dataset of patients \code{tableau} generated with the function
\code{\link{gen_patients_multinomTOP}} or with a similar structure. The function runs for 1 trial,
and given the data and the threshold, it simulates the ongoing of the trial and return the final result.
The TOP is a single-arm design. The code is from Ruitao Lin.
}
\examples{
# Run a trial of 80 patients with 3 interim analyses, 5 patients per month and 6 months
# of observation for efficacy and toxicity
test_pat <- gen_patients_multinomTOP(n_sim = 1L, ana_inter = rep(20, 4),
                                     interpatient = 6, max_teff = 180, max_ttox = 180,
                                     multinom_ttt = c(.1, .05, .15, .7))
realisation_essai(ana_inter = rep(20, 4),
                  tableau = test_pat[[1]],
                  phitox = c(.15, .25), prior_eff = .15, prior_tox = .25,
                  C_ = .62, gamm = .99,
                  interpatient = 6, max_teff = 180, max_ttox = 180)
}
\seealso{
\code{\link{gen_patients_multinomTOP}}
}
