% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opcharac_efftoxTOP.R
\name{simu_essais}
\alias{simu_essais}
\title{Simulate single-arm trials with the TOP design}
\usage{
simu_essais(
  alpha = 0.1,
  bonf = FALSE,
  nsim_oc = 10000,
  nsim_essai = 10000,
  ana_inter,
  ana_inter_tox = NULL,
  p_n,
  p_a,
  p_reel,
  liste_tableaux = NULL,
  phi = NULL,
  prior = NULL,
  cut_seq = seq(0.5, 0.95, by = 0.005),
  power_seq = seq(0, 1, by = 0.01),
  interpatient,
  max_teff,
  max_ttox,
  seed_simu = 1993,
  seed_cut = 1024,
  mat_beta_xi = matrix(c(1, 1, 0, 0, 0, 1, 0, 1), nrow = 2, byrow = TRUE),
  stock_matrice = FALSE
)
}
\arguments{
\item{alpha}{Maximal type I error rate desired.}

\item{bonf}{In case of multi-arm trial, TRUE indicates a Bonferroni correction to optimize Cn and FALSE no correction. For
now, TOP is onyl in single-arm so not useful yet.}

\item{nsim_oc, nsim_essai}{Number of simulated trials to optimize the threshold Cn and determine the operating characteristics.}

\item{ana_inter}{A vector of the supplementary patients recruited at each interim analysis compared with the previous one.}

\item{ana_inter_tox}{Vector of the number of additional patients at each interim analysis for toxicity. If analyses for effficacy and toxicity occur at the same
number of patients, set it to NULL.}

\item{p_n, p_a, p_reel}{Vectors of length 4 giving the multinomial law of probability for the unefficacy/toxicity hypothesis,
the efficacy/safety hypothesis and the simulated outcomes in the trials (\code{p_reel} = "real probabilities")}

\item{liste_tableaux}{List of datasets or NULL if you want the function to generate them.}

\item{phi}{Theoretical thresholds to compare values of efficacy and toxicity.}

\item{prior}{The prior for efficacy and toxicity.}

\item{cut_seq, power_seq}{Vectors giving the screening values for both tuning parameters of Cn. If only 1 value supplied in each
argument, no grid search is performed and these values become the 2 optimized parameters.}

\item{interpatient}{Mean time between 2 patients.}

\item{max_teff, max_ttox}{Length of the observation window for efficacy and toxicity.}

\item{seed_simu, seed_cut}{The seeds used to make reproducible the use of \code{rmultinom} (in the simulation of the trials'
datasets to determine the operating characteristics and to optimize the threshold Cn).}

\item{mat_beta_xi}{The link matrix between outcomes of the multinomial distribution and endpoints (careful, in the second row
1 is actually for no toxicity /!\).}

\item{stock_matrice}{If TRUE, return the operating characteristics and the table of the decision in each trial, and
if FALSE, return only the operating characteristics.}
}
\value{
The operating characteristics of the design applied to scenario defined in \code{p_reel} or in \code{liste_patients}.
}
\description{
Either run the trials in \code{liste_tableaux} and give the results of all trials, or
simulate them and give the results. If \code{liste_tableaux} is NULL, you must supply the
arguments of the function \code{gen_patients_multinomTOP} to allow it to generate the data.
}
\examples{
simu_essais(alpha = .05,
            ana_inter = c(30, 51),
            p_n = c(.1, .05, .15, .7),
            p_a = c(.12, .18, .03, .67),
            p_reel = c(.12, .18, .03, .67),
            interpatient = 6,
            max_teff = 180,
            max_ttox = 180)
}
